pyproject.toml
[project]
name = "smart-interview-guideline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "email-validator>=2.0.0",
    "fastapi>=0.129.0",
    "greenlet>=3.1.1",
    "gunicorn>=25.1.0",
    "openai>=2.20.0",
    "bcrypt>=4.0.0",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "pypdf>=5.0.0",
    "httpx>=0.27.0",
    "python-docx>=1.0.0",
    "sqlmodel>=0.0.33",
    "uvicorn>=0.40.0",
]
